chore: rename cronicle plugins to xyops
