+++ tests/movie_pipeline/jobs/test_process_movie.py
@@ -9,7 +9,7 @@ from unittest.mock import patch
 
 import movie_pipeline.jobs.main as job
 
-from ..concerns import copy_files, create_output_movies_directories, get_base_cronicle_json_input, get_movie_edl_file_content, get_output_movies_directories, get_serie_edl_file_content
+from ..concerns import copy_files, create_output_movies_directories, get_base_xyops_json_input, get_movie_edl_file_content, get_output_movies_directories, get_serie_edl_file_content
 
 
 class ProcessMovieTest(unittest.TestCase):
@@ -35,16 +35,16 @@ class ProcessMovieTest(unittest.TestCase):
         logo_dir_path = self.input_dir_path / 'logo'
         logo_dir_path.mkdir(parents=True)
 
-        self.cronicle_json_input = get_base_cronicle_json_input()
+        self.xyops_json_input = get_base_xyops_json_input()
 
     @patch('sys.stdout', new_callable=StringIO)
     def test_log_progress_of_process_movie(self, mock_stdout):
         edl_path = self.video_path.with_suffix('.mp4.yml')
         edl_path.write_text(get_movie_edl_file_content(), encoding='utf-8')
 
-        self.cronicle_json_input["params"] = {'file_path': str(self.video_path.absolute()), 'edl_ext': '.yml'}
+        self.xyops_json_input["params"] = {'file_path': str(self.video_path.absolute()), 'edl_ext': '.yml'}
 
-        with patch.object(sys, 'stdin', StringIO(json.dumps(self.cronicle_json_input))):
+        with patch.object(sys, 'stdin', StringIO(json.dumps(self.xyops_json_input))):
             job.process_movie(self.config_path)
 
         output = mock_stdout.getvalue()
@@ -52,7 +52,7 @@ class ProcessMovieTest(unittest.TestCase):
         self.assertRegex(output, re.compile(r'"perf": {"ProcessStep": [\d.]+, "BackupStep": [\d.]+}'))
         self.assertRegex(output, '{"xy": 1, "code": 0}')
 
-    @patch('movie_pipeline.services.movie_file_processor.runner.cronicle.cronicle_runner.requests.post')
+    @patch('movie_pipeline.services.movie_file_processor.runner.xyops.xyops_runner.requests.post')
     @patch('sys.stdout', new_callable=StringIO)
     def test_log_progress_of_process_directory(self, mock_stdout, mock_post):
         edl_path = self.video_path.with_suffix('.mp4.yml')
@@ -64,14 +64,14 @@ class ProcessMovieTest(unittest.TestCase):
         mock_post.return_value.status_code = 200
         mock_post.return_value.json.side_effect = [{'code': 0, 'ids': ['23f5c37f']}, {'code': 0, 'ids': [], 'queue': 1}]
 
-        self.cronicle_json_input["params"] = {'api_key': 'CRONICLE_API_KEY', 'folder_path': str(self.input_dir_path.absolute()), 'edl_ext': '.yml'}
+        self.xyops_json_input["params"] = {'api_key': 'CRONICLE_API_KEY', 'folder_path': str(self.input_dir_path.absolute()), 'edl_ext': '.yml'}
 
-        with patch.object(sys, 'stdin', StringIO(json.dumps(self.cronicle_json_input))):
+        with patch.object(sys, 'stdin', StringIO(json.dumps(self.xyops_json_input))):
             job.process_directory(self.config_path)
 
         common_expected_post_args = ['http://localhost:3012/api/app/run_event/v1']
         common_expected_post_kwargs = {
-            'headers': {'X-API-Key': self.cronicle_json_input['params']['api_key']},
+            'headers': {'X-API-Key': self.xyops_json_input['params']['api_key']},
             'json': {'title': 'Process Movie'}
         }
